Cloud-Valet/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
alembic
python-dotenv
psycopg2-binary
passlib
jinja2
httpx
pytest
